// src/api/generated/schemas/registerIpRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * cid-api
 * LDAP 로그인 + JWT 인가 API
 * OpenAPI spec version: v1
 */

export interface RegisterIpRequest {
  /**
   * IP 주소 (IPv4 or IPv6)
   * @minLength 0
   * @maxLength 45
   */
  ipAddr: string;
  /**
   * IP 타입 (REAL/VIP/ADMIN/NAS/PUBLIC/PRIVATE)
   * @minLength 0
   * @maxLength 20
   */
  ipTpCd: string;
  /**
   * 호스트 명
   * @minLength 0
   * @maxLength 100
   */
  hostNm?: string;
  /**
   * IP 설명
   * @minLength 0
   * @maxLength 200
   */
  ipDescp?: string;
  /**
   * MAC 주소
   * @minLength 0
   * @maxLength 17
   */
  macAddr?: string;
  /**
   * DNS 명
   * @minLength 0
   * @maxLength 200
   */
  dnsNm?: string;
  /** CI 고유 ID */
  ciId?: number;
  /** SUBNET ID */
  subnetId?: number;
}